langchain_tools/pyproject.toml
[tool.poetry]
name = "langchain-tools"
version = "0.1.0"
description = "A LangChain-based CLI tool with multiple tools and LaunchDarkly integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = ">=0.2"
langchain-community = ">=0.2"
langchain-openai = ">=0.1"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
rich = "^13.0.0"
click = "^8.1.0"
launchdarkly-server-sdk = "^9.12.0"
launchdarkly-server-sdk-ai = "^0.10.0"
launchdarkly-observability = "^0.1.0"
opentelemetry-instrumentation-openai = ">=0.40.14"
opentelemetry-instrumentation-langchain = "0.40.14"
tiktoken = ">=0.6"
watchdog = "^3.0.0"
duckduckgo-search = ">=6"
numexpr = "^2.8.7"
pydantic = ">=2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.cli:main"
watch = "src.dev:main"